9/requirements.txt
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
billiard==4.2.0
cachetools==5.3.3
celery==5.3.6
chardet==5.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
distlib==0.3.8
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.13.3
flower==2.0.1
greenlet==3.0.3
h11==0.14.0
humanize==4.9.0
idna==3.6
kombu==5.3.6
Mako==1.3.2
MarkupSafe==2.1.5
packaging==24.0
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pyproject-api==1.6.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.36.3
tesseract==0.1.3
tomli==2.0.1
tornado==6.4
tox==4.14.2
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.29.0
vine==5.1.0
virtualenv==20.25.1
wcwidth==0.2.13
